% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{jobr_benchmark_jobs}
\alias{jobr_benchmark_jobs}
\title{Build a benchmark jobs data frame}
\usage{
jobr_benchmark_jobs(n = 240, seconds = 1)
}
\arguments{
\item{n}{Number of jobs.}

\item{seconds}{CPU seconds each job should take.}
}
\value{
A data frame suitable for passing to \code{\link[=host_new]{host_new()}}.
}
\description{
Build a benchmark jobs data frame
}
\examples{
head(jobr_benchmark_jobs(10, seconds = 0.5))
}
