% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spool.R
\name{jobr_spool_clear}
\alias{jobr_spool_clear}
\title{Throw away spooled results}
\usage{
jobr_spool_clear(jobset = NULL, dir = spool_dir())
}
\arguments{
\item{jobset}{Restrict to one jobset, or NULL for all.}

\item{dir}{Spool directory.}
}
\value{
The number removed, invisibly.
}
\description{
Throw away spooled results
}
