% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ledger.R
\name{ledger_open}
\alias{ledger_open}
\title{Open (or create) a ledger}
\usage{
ledger_open(path)
}
\arguments{
\item{path}{Path to the ledger file. Created if absent.}
}
\value{
A \code{jobr_ledger} object: a list of the path and the replayed state.
}
\description{
Open (or create) a ledger
}
