% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{manifest_read}
\alias{manifest_read}
\title{Read a project manifest}
\usage{
manifest_read(dir)
}
\arguments{
\item{dir}{Project directory containing \code{jobR.dcf}.}
}
\value{
A list with \code{project}, \code{entrypoint}, \code{files}, and \code{lockfile}.
}
\description{
Read a project manifest
}
