% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrol.R
\name{passphrase_new}
\alias{passphrase_new}
\title{Generate a human-transferable passphrase}
\usage{
passphrase_new(n_words = 4)
}
\arguments{
\item{n_words}{Number of words. Four words from the EFF long list carry
roughly 51 bits, which is ample for a credential that is also rate-limited
by a human typing it.}
}
\value{
A single dash-separated string.
}
\description{
Generate a human-transferable passphrase
}
