% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrol.R
\name{passphrase_wordlist}
\alias{passphrase_wordlist}
\title{The passphrase wordlist}
\usage{
passphrase_wordlist()
}
\value{
A character vector of candidate words.
}
\description{
The passphrase wordlist
}
\keyword{internal}
