% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrol.R
\name{passphrase_words}
\alias{passphrase_words}
\title{Split a passphrase into its words}
\usage{
passphrase_words(phrase)
}
\arguments{
\item{phrase}{A dash-separated passphrase.}
}
\value{
A character vector of words, lowercased and trimmed.
}
\description{
Split a passphrase into its words
}
