% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobR-package.R
\docType{data}
\name{passphraseWords}
\alias{passphraseWords}
\title{Words used to build human-transferable passphrases}
\format{
A data frame with \code{id} and \code{word} columns.
}
\source{
\url{https://www.eff.org/files/2016/07/18/eff_large_wordlist.txt}
}
\usage{
passphraseWords
}
\description{
The EFF long wordlist, chosen for words that are unambiguous when spoken
aloud and hard to mistype.
}
\keyword{datasets}
