% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{r_version_string}
\alias{r_version_string}
\title{This machine's R version as a short string}
\usage{
r_version_string()
}
\value{
For example \code{"4.5.3"}.
}
\description{
This machine's R version as a short string
}
