% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_path}
\alias{repo_path}
\title{Path to a host's package repository}
\usage{
repo_path(work_dir)
}
\arguments{
\item{work_dir}{The host's working directory.}
}
\value{
The repository root.
}
\description{
Path to a host's package repository
}
