% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_url}
\alias{repo_url}
\title{A \verb{file://} URL R will accept for a local repository}
\usage{
repo_url(dir)
}
\arguments{
\item{dir}{Local repository root.}
}
\value{
A URL string.
}
\description{
A \verb{file://} URL R will accept for a local repository
}
