% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spool.R
\name{spool_drop}
\alias{spool_drop}
\title{Discard spooled results}
\usage{
spool_drop(paths)
}
\arguments{
\item{paths}{Paths from \code{\link[=spool_list]{spool_list()}}.}
}
\value{
The number removed, invisibly.
}
\description{
Discard spooled results
}
