% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spool.R
\name{spool_write}
\alias{spool_write}
\title{Set aside results that could not be delivered}
\usage{
spool_write(jobset, chunk, results, dir = spool_dir())
}
\arguments{
\item{jobset}{Jobset identifier.}

\item{chunk}{Chunk index.}

\item{results}{The per-job results for that chunk.}

\item{dir}{Spool directory.}
}
\value{
The path written, invisibly.
}
\description{
Set aside results that could not be delivered
}
