% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrol.R
\name{token_new}
\alias{token_new}
\title{Generate an opaque session token}
\usage{
token_new(bytes = 16)
}
\arguments{
\item{bytes}{Number of random bytes.}
}
\value{
A hex string.
}
\description{
Generate an opaque session token
}
